+++ cli/src/ui/components/StatusIndicator.tsx
@@ -3,12 +3,9 @@
  * Shows status text on the left (e.g., "Thinking...") and available hotkeys on the right
  */
 
-import React, { useCallback } from "react"
+import React from "react"
 import { Box, Text } from "ink"
-import { useKeyboard } from "../../state/hooks/useKeyboard.js"
-import { HOTKEYS } from "../../constants/keyboard/hotkeys.js"
 import { useHotkeys } from "../../state/hooks/useHotkeys.js"
-import { useWebviewMessage } from "../../state/hooks/useWebviewMessage.js"
 import { useTheme } from "../../state/hooks/useTheme.js"
 import { HotkeyBadge } from "./HotkeyBadge.js"
 import { useAtomValue } from "jotai"
@@ -30,62 +27,12 @@ export interface StatusIndicatorProps {
  * - Shows approval hotkeys when approval is pending
  * - Shows navigation hotkeys when followup suggestions are visible
  * - Shows general command hints when idle
- * - Handles Ctrl+X / Cmd+X to cancel tasks
  */
 export const StatusIndicator: React.FC<StatusIndicatorProps> = ({ disabled = false }) => {
 	const theme = useTheme()
 	const { hotkeys, shouldShow } = useHotkeys()
-	const { cancelTask, resumeTask } = useWebviewMessage()
 	const isStreaming = useAtomValue(isStreamingAtom)
 	const hasResumeTask = useAtomValue(hasResumeTaskAtom)
-
-	// Handle Ctrl+X / Cmd+X to cancel when streaming
-	const handleCancel = useCallback(async () => {
-		if (isStreaming && !disabled) {
-			try {
-				await cancelTask()
-			} catch (error) {
-				// Silently handle task abortion errors as they're expected
-				const isTaskAbortError =
-					error instanceof Error &&
-					error.message &&
-					error.message.includes("task") &&
-					error.message.includes("aborted")
-
-				if (!isTaskAbortError) {
-					console.error("Failed to cancel task:", error)
-				}
-			}
-		}
-	}, [isStreaming, disabled, cancelTask])
-
-	// Handle Ctrl+R / Cmd+R to resume task
-	const handleResume = useCallback(async () => {
-		if (hasResumeTask && !disabled) {
-			try {
-				await resumeTask()
-			} catch (error) {
-				console.error("Failed to resume task:", error)
-			}
-		}
-	}, [hasResumeTask, disabled, resumeTask])
-
-	// Listen for Ctrl+X to cancel and Ctrl+R to resume
-	useKeyboard(
-		{
-			hotkeys: [
-				{
-					hotkey: HOTKEYS.CANCEL_TASK,
-					handler: handleCancel,
-				},
-				{
-					hotkey: HOTKEYS.RESUME_TASK,
-					handler: handleResume,
-				},
-			],
-		},
-		{ active: !disabled && (isStreaming || hasResumeTask) },
-	)
 
 	// Don't render if no hotkeys to show or disabled
 	if (!shouldShow || disabled) {
